Update webpack.config & package.json to reflect local dev & heroku prod environments
- seperated webpack.config into two files - webpack.dev.config & webpack.prod.config
  dev is specific to local dev env, prod applies to heroku staging and prod
- implemented package.json scripts / renamed existing to reflect what they do and what environments they're relevant to:
-- build-dev: builds the client and server files for local dev env and watches for changes to them
-- dev-server: runs the local dev server
-- dev-database: runs the local dev database
-- client-dev-server: runs local client dev server, only serves client files, hot-updates the browser on file change
-- build-prod: builds the client and server files for heroku prod env
-- prod-server: runs the heroku dev server
-- heroku-postbuild: special heroku hook for running commands after a heroku deployment
