+++ package.json
@@ -1,15 +1,18 @@
 {
-  "name": "express-react-redux-starter",
+  "name": "hr-xmen-hungry-photo",
   "version": "1.1.0",
   "description": "Starter for Express, React, Redux, SCSS applications",
   "scripts": {
-    "dev": "webpack-dev-server --watch --colors",
-    "build": "rimraf dist && NODE_ENV=production webpack --colors",
-    "server": "NODE_ENV=production PORT=8080 nodemon ./server/server.js",
+    "build-dev": "rimraf dist && webpack --watch  --colors --config webpack.dev.config",
+    "dev-server": "NODE_ENV=development PORT=8080 nodemon ./server/server.js",
+    "dev-database": "mongod --dbpath ./db",
+    "client-dev-server": "CLIENT_DEV=true CLIENT_DEV_PORT=3000 webpack-dev-server --watch --colors --config webpack.dev.config --port 3000",
+    "build-prod": "rimraf dist && webpack --colors -p --config webpack.prod.config",
+    "prod-server":"node ./server/server.js",
+    "heroku-postbuild": "npm run build-prod",
     "test": "mocha --compilers js:babel-core/register --require ./test/test_helper.js --recursive ./test",
     "test:watch": "npm run test -- --watch",
-    "lint": "eslint src test webpack",
-    "database": "mongod --dbpath ./db"
+    "lint": "eslint src test webpack"
   },
   "keywords": [
     "ExpressJS",
@@ -29,7 +32,7 @@
     "type": "git",
     "url": ""
   },
-  "author": "Alexi Taylor",
+  "author": "hr-xmen",
   "license": "MIT",
   "devDependencies": {
     "babel-core": "^6.8.0",
@@ -45,8 +48,6 @@
     "eslint-plugin-import": "^1.8.0",
     "eslint-plugin-jsx-a11y": "^1.2.0",
     "eslint-plugin-react": "^5.1.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "html-webpack-plugin": "^2.16.1",
     "jquery": "^2.2.3",
     "jsdom": "^9.0.0",
     "mocha": "^2.4.5",
@@ -56,7 +57,6 @@
     "sass-loader": "^3.2.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.0",
     "webpack-dev-server": "^1.14.1",
     "webpack-hot-middleware": "^2.18.0"
   },
@@ -70,8 +70,10 @@
     "connect-ensure-login": "^0.1.1",
     "cookie-parser": "^1.4.3",
     "express": "^4.13.4",
-    "flux": "^3.1.2",
     "express-session": "^1.15.2",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "flux": "^3.1.2",
+    "html-webpack-plugin": "^2.16.1",
     "material-ui": "^0.18.0",
     "mongoose": "^4.9.8",
     "morgan": "^1.8.1",
@@ -90,6 +92,7 @@
     "request": "^2.81.0",
     "request-promise": "^4.2.1",
     "underscore": "^1.8.3",
+    "webpack": "^1.13.0",
     "yelp-api-v3": "^1.5.0",
     "yummly": "^0.2.4"
   }
